Add files via upload
This is the Arduino portion of the IP address display project. The code here reads the incoming serial data from the Jetson and then prints it out to the LCD. The code does work, however it does seem to randomly cut off what's being displayed sometimes and then returns the IP again. Could be a timing issue in the Arduino file or in the shell script file. In any case, the IP is displayed for a long enough period of time that it can be used.